Backend/pyproject.toml
[tool.poetry]
name = "testwise-backend"
version = "0.1.0"
description = "Backend for Educational Platform"
authors = ["Artyom Pereiatentcev <pereiatentcev@gmail.com>"]
packages = [{include = "src", from = "."}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.35"
aiosqlite = "^0.20.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.30.0"
alembic = "^1.13.0"
minio = "^7.2.0"
pydantic = { version = "^2.0", extras = ["email"] }
pydantic-settings = "^2.5.2"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
bcrypt = "4.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiofiles = "^23.1.0"
python-multipart = "^0.0.6"
requests = "^2.32.5"
aiohttp = "^3.10.0"
redis = {extras = ["hiredis"], version = "^5.0.0"}
pillow = "^12.0.0"
pdf2image = "^1.17.0"
isort = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
black = "^25.9.0"
ruff = "^0.14.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"